add/ test to check values ​​of the DB arent inconsistent I have created a series of checks to ensure that the database does not contain inconsistencies before allowing the use of the seed:
-No range outages or creation of extra categories (values) due to writing errors

-That the veracity of both exclusivity and the possibility of sharing the game be verified when the conditions "exclusiveness:" true and "sharedWhit": !=[] coexist

And the same when the opposite happens. (false, and ['switch', 'play'] )

-If the name of the console is found in the array of the "sharedWhit" property, it checks if the repetition error is in the "console" property (repetition of the game-console binomial) or in the "sharedWhit" property ( change the name to the corresponding console) by observing the repetition in the "name" property

-That the trailers are not embed

-That the strings of the images begin with ../assets/ and end with .jpg

In case of passing all the tests, the seed updates the database, if not,
you just have to follow the instructions (Much more declarative and
brief than this) of the console to replace the errors.
