+++ package.json
@@ -7,7 +7,8 @@
     "seed": "node seed/seed.js",
     "test": "jest",
     "start": "node server.js",
-    "dev": "nodemon server.js"
+    "dev": "nodemon server.js",
+    "testDB": "node seed/testValuesDB.js"
   },
   "repository": {
     "type": "git",
@@ -22,10 +23,12 @@
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
+    "emojis-list": "^3.0.0",
     "express": "^4.18.2",
     "express-openid-connect": "^2.12.0",
     "mongoose": "^6.9.0",
     "node": "^18.13.0",
+    "node-emoji": "^1.11.0",
     "node-fetch": "^3.3.0",
     "openssl": "^2.0.0",
     "request": "^2.88.2"
