+++ seed/games.js
@@ -1,0 +1,4028 @@
+//creamos los provedores y los productos
+const providers = [
+
+    {
+        "company_name": "Gatos0",
+        "CIF": "F7DS9F98S",
+        "address": "Calle del persa",
+    },
+    {
+        "company_name": "Gatos1",
+        "CIF": "F9S7D98SD",
+        "address": "Calle del egipcio",
+    },
+    {
+        "company_name": "Gatos2",
+        "CIF": "F9S7D9FSD",
+        "address": "Calle de mala gata",
+    },
+    {
+        "company_name": "Gatos3",
+        "CIF": "79FSD98",
+        "address": "Calle de la human-cat",
+    },
+    {
+        "company_name": "Gatos4",
+        "CIF": "GSF799GS9D",
+        "address": "Calle del gatito",
+    },
+    {
+        "company_name": "Gatos5",
+        "CIF": "G9S7DG9SD",
+        "address": "Calle de todas las razas",
+    },
+    {
+        "company_name": "Gatos6",
+        "CIF": "GSD98GSD",
+        "address": "Calle de gatos",
+    },
+    {
+        "company_name": "Gatos7",
+        "CIF": "G8SD6D69GSD",
+        "address": "Calle del lince",
+    }, {
+        "company_name": "Gatos8",
+        "CIF": "G87SF9S",
+        "address": "Calle del callejero",
+    },
+    {
+        "company_name": "Gatos9",
+        "CIF": "79SFDGSD9",
+        "address": "Calle del minino",
+    }
+
+]
+
+const products = [
+    {
+        "image": "../assets/spiderMan.jpg",
+        "name": "Spider-Man Miles Morales",
+        "opinion": 4.7,
+        "price": 60.25,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Insomniac Games"],
+        "genre": ["Adventure"],
+        "trailer": "https://www.youtube.com/embed/Q3kfF3XNzw8",
+    },
+    {
+        "image": "../assets/horizon.jpg",
+        "name": "Horizon",
+        "opinion": 4.7,
+        "price": 79.98,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Guerrilla Games"],
+        "genre": ["Rol", "Adventure", "Open World"],
+        "trailer": "https://www.youtube.com/embed/p9a2SlsbV3w",
+    },
+    {
+        "image": "../assets/ratchetClank.jpg",
+        "name": "Ratchet Clank",
+        "opinion": 4.8,
+        "price": 79.19,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Insomniac Games"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embed/IkPWYH403tU",
+    },
+    {
+        "image": "../assets/granTurismo.jpg",
+        "name": "Gran Turismo",
+        "opinion": 4.5,
+        "price": 46.98,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Polyphony Digital"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embed/1tBUsXIkG1A",
+    },
+    {
+        "image": "../assets/callOfDuty.jpg",
+        "name": "Call of Duty: Modern Warfare II",
+        "opinion": 4.6,
+        "price": 79.99,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Infinity Ward"],
+        "genre": ["Shoter"],
+        "trailer": "https://www.youtube.com/embed/T57a2wE_hpk",
+    },
+    {
+        "image": "../assets/sonicFrontiers.jpg",
+        "name": "Sonic Frontiers",
+        "opinion": 4.6,
+        "price": 39.99,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Sonic Team"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embed/30-rKqe1Ys8",
+    },
+    {
+        "image": "../assets/minecraftLegends.jpg",
+        "name": "Minecraft Leyends",
+        "opinion": 4.2,
+        "price": 49.99,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Nintendo Switch"],
+        "developers": ["Mojang Studios"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embed/L1o-3x0KNVc",
+    },
+    {
+        "image": "../assets/nickelodeonAllStarBrawl.jpg",
+        "name": "Nickelodeon All Star Brawl",
+        "opinion": 4.3,
+        "price": 32.67,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Nintendo Switch"],
+        "developers": ["Ludosity"],
+        "genre": ["Fight"],
+        "trailer": "https://www.youtube.com/embed/xmXEt9lgpz0",
+    },
+    {
+        "image": "../assets/metroExodus.jpg",
+        "name": "Metro Exodus",
+        "opinion": 4.3,
+        "price": 29.70,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["4A Games"],
+        "genre": ["Shoter", "Horror"],
+        "trailer": "https://www.youtube.com/embed/c_YddLpfD5o",
+    },
+    {
+        "image": "../assets/likeADragonIshin.jpg",
+        "name": "Like A Dragon: ISHIN",
+        "opinion": 4.7,
+        "price": 62.78,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Ryu ga Gotoku Studio"],
+        "genre": ["Fight", "Adventure", "Open World"],
+        "trailer": "https://www.youtube.com/embed/xT-Z_SgPC2w",
+    },
+    {
+        "image": "../assets/howartsLegacy.jpg",
+        "name": "Howarts Legacy",
+        "opinion": 4.8,
+        "price": 75.27,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Nintendo Switch"],
+        "developers": ["Portkey Games"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/S6GTl_vPRvU",
+    },
+    {
+        "image": "../assets/forspoken.jpg",
+        "name": "Forspoken",
+        "opinion": 4.3,
+        "price": 70.05,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Luminous Productions"],
+        "genre": ["Rol"],
+        "trailer": "https://www.youtube.com/embed/9kMXdmswTFQ",
+    },
+    {
+        "image": "../assets/wolongFallenDinasty.jpg",
+        "name": "WO LONG Fallen Dinasty",
+        "opinion": 3.8,
+        "price": 72.99,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Team Ninja"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/xlmdeQFyqC0",
+    },
+    {
+        "image": "../assets/scarsAbove.jpg",
+        "name": "SCARS above",
+        "opinion": 4.2,
+        "price": 42.84,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Koch Media GmbH"],
+        "genre": ["Shoter", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/Mvj5apgk7kI",
+    },
+    {
+        "image": "../assets/aeternaNoctis.jpg",
+        "name": "Aeterna Noctis",
+        "opinion": 4.4,
+        "price": 40.31,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Aeternum Game Studio"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embed/6ob06okmZAQ",
+    },
+    {
+        "image": "../assets/valkiyrieElysium.jpg",
+        "name": "Valkyrie Elysium",
+        "opinion": 4.5,
+        "price": 72.9,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Team Ninja"],
+        "genre": ["Rol", "Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/0QWxP3iFmvs",
+    },
+    {
+        "image": "../assets/astronite.jpg",
+        "name": "Astronite",
+        "opinion": 4.7,
+        "price": 24.99,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Dume Games Studio"],
+        "genre": ["Platforms", "Adventure", "Shoter"],
+        "trailer": "https://www.youtube.com/embed/xlfcwc-OTd4",
+    },
+    {
+        "image": "../assets/relayer.jpg",
+        "name": "Relayer",
+        "opinion": 4.6,
+        "price": 59.90,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Kadokawa Games"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/59u6R8DUz1E",
+    },
+    {
+        "image": "../assets/evilWest.jpg",
+        "name": "Evil West",
+        "opinion": 4.0,
+        "price": 36.58,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Flying Wild Hog"],
+        "genre": ["Shoter"],
+        "trailer": "https://www.youtube.com/embed/2hP6IVcHRnQ",
+    },
+    {
+        "image": "../assets/aPlaguetaleRequiem.jpg",
+        "name": "A Plague Tale: Requiem",
+        "opinion": 4.7,
+        "price": 63.04,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Asobo Studio"],
+        "genre": ["Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/NAuzkT5A4xM",
+    },
+    {
+        "image": "../assets/shadowWarrior.jpg",
+        "name": "Shadow Warrior",
+        "opinion": 3.7,
+        "price": 40.20,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Flying Wild Hog"],
+        "genre": ["Shoter", "Horror"],
+        "trailer": "https://www.youtube.com/embed/zfcK2vlX95U",
+    },
+    {
+        "image": "../assets/fairyFencerFRefrainChord.jpg",
+        "name": "Fairy Fencer F: Refrain Chord",
+        "opinion": 4.1,
+        "price": 59.99,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Sting Entertainment", "Compile Heart", "Idea Factory"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/E0in01DbFHE",
+    },
+    {
+        "image": "../assets/neptuniaSisterVsSister.jpg",
+        "name": "Neptunia: Sister Vs Sister",
+        "opinion": 4.8,
+        "price": 60.32,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Compile Heart", "Idea Factory"],
+        "genre": ["Rol", "Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embed/aSjeDksK6nw",
+    },
+    {
+        "image": "../assets/neptuniaReverse.jpg",
+        "name": "Neptunia: Re Verse",
+        "opinion": 4.3,
+        "price": 50.39,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Compile Heart", "Idea Factory", "Felistella"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/ixug9quQJaM",
+    },
+    {
+        "image": "../assets/unchartedElLegadoPerdido.jpg",
+        "name": "UNCHARTED: El legado Perdido",
+        "opinion": 4.7,
+        "price": 34.47,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Naughty Dog"],
+        "genre": ["Adventure", "Shoter"],
+        "trailer": "https://www.youtube.com/embed/Q7_7Y471qHk",
+    },
+    {
+        "image": "../assets/sackBoyABigAventure.jpg",
+        "name": "SACKBOY: A Big Aventure",
+        "opinion": 4.6,
+        "price": 69.75,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Sumo Digital", "xdev software gmbh"],
+        "genre": ["Platforms", "Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embed/MEoy6gqOsRE",
+    },
+    {
+        "image": "../assets/finalFantasyVII.jpg",
+        "name": "FINAL FANTASY VII",
+        "opinion": 4.7,
+        "price": 54.90,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Square Enix"],
+        "genre": ["Fight", "Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/Ge73iBqc7o8",
+    },
+    {
+        "image": "../assets/goodByeWorld.jpg",
+        "name": "Good by world",
+        "opinion": 4.1,
+        "price": 39.99,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["yo fujii", "isolation studio"],
+        "genre": ["Platforms", "Classic"],
+        "trailer": "https://www.youtube.com/embed/Gc3eRKsSFxg",
+    },
+    {
+        "image": "../assets/8DoorsArumsAfterlifeAdvenure.jpg",
+        "name": "8 DOORS: Arum`s after life advenure",
+        "opinion": 4.8,
+        "price": 29.99,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Metroidvania"],
+        "genre": ["Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embed/RbAPvuYb7FA",
+    },
+    {
+        "image": "../assets/deliverUsTheMoon.jpg",
+        "name": "DELIVER US THE MOON",
+        "opinion": 4.5,
+        "price": 24.99,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["KeokeN Interactive"],
+        "genre": ["Strategy", "Adventure", "Platforms"],
+        "trailer": "https://www.youtube.com/embed/RB17XR-i4hM",
+    },
+    {
+        "image": "../assets/deathsDoor.jpg",
+        "name": "DEATH`S DOOR",
+        "opinion": 4.7,
+        "price": 29.99,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Acid Nerve"],
+        "genre": ["Rol", "Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embed/MxzoW_fUx_8",
+    },
+    {
+        "image": "../assets/ghostBusterSpiritsUnleashed.jpg",
+        "name": "GHOST BUSTER: Spirits Unleashed",
+        "opinion": 3.3,
+        "price": 40.31,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["IllFonic"],
+        "genre": ["Shoter"],
+        "trailer": "https://www.youtube.com/embed/OYWPBzz_wmI",
+    },
+    {
+        "image": "../assets/vampireTheMasquerade.jpg",
+        "name": "VAMPIRE: The masquerade",
+        "opinion": 4.1,
+        "price": 51.33,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["HardSuit Labs", "Paradox Interactive"],
+        "genre": ["Rol", "Horror", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/TasT4b1qweA",
+    },
+    {
+        "image": "../assets/apsulovEndOfGods.jpg",
+        "name": "APSULOV: end of gods",
+        "opinion": 4.3,
+        "price": 34.36,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Angry Demon Studio"],
+        "genre": ["Horror", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/IuK0JQopc7c",
+    },
+    {
+        "image": "../assets/theCalistoProtocol.jpg",
+        "name": "The Calisto Protocols",
+        "opinion": 4.3,
+        "price": 69.99,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["PUBG Studios", "Striking distance studios"],
+        "genre": ["Horror", "Shoter", "Fight"],
+        "trailer": "https://www.youtube.com/embed/BgiBtSH4ZyE",
+    },
+    {
+        "image": "../assets/marthaIsDead.jpg",
+        "name": "Martha is Dead",
+        "opinion": 4.4,
+        "price": 29.90,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["LKA"],
+        "genre": ["Horror", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/QpX6RSNQodQ",
+    },
+    {
+        "image": "../assets/doNotOpen.jpg",
+        "name": "Do not Open",
+        "opinion": 4.8,
+        "price": 30.12,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["NOX NOCTIS", "Quasar Dynamics"],
+        "genre": ["Horror", "Strategy"],
+        "trailer": "https://www.youtube.com/embed/SfHpo5WcQEI",
+    },
+    {
+        "image": "../assets/theWine.jpg",
+        "name": "Horror Tales: The Wine - Announcement",
+        "opinion": 4.8,
+        "price": 26.87,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Avance"],
+        "genre": ["Horror", "Fight"],
+        "trailer": "https://www.youtube.com/embed/ckGjOtdF7_s",
+    },
+    {
+        "image": "../assets/fiveNightsAtFreddysSecurityBreach.jpg",
+        "name": "Five Nights at Freddy's: Security Breach",
+        "opinion": 4.5,
+        "price": 42.94,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Steel Wool Studios"],
+        "genre": ["Adventure", "Horror"],
+        "trailer": "https://www.youtube.com/embed/g8hHZr3GpWI",
+    },
+    {
+        "image": "../assets/stray.jpg",
+        "name": "STRAY",
+        "opinion": 4.7,
+        "price": 40.23,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Steel Wool Studios"],
+        "genre": ["Horror", "Adventure"],
+        "trailer": "https://www.youtube.com/embedvBDQEkfAcnc",
+    },
+    {
+        "image": "../assets/kenaBridgeOfSpirits.jpg",
+        "name": "KENA: Bridge of the spirits",
+        "opinion": 4.7,
+        "price": 53.23,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Ember Lab"],
+        "genre": ["Adventure", "Open World"],
+        "trailer": "https://www.youtube.com/embed/pWh5388AEHw",
+    },
+    {
+        "image": "../assets/steelRising.jpg",
+        "name": "STEEL RISING",
+        "opinion": 4.5,
+        "price": 59.62,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Spiders"],
+        "genre": ["Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embed/DlZwlT08mf0",
+    },
+    {
+        "image": "../assets/gothamKnights.jpg",
+        "name": "GOTHAM KNIGTHS",
+        "opinion": 4.1,
+        "price": 75.45,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["QLOC", "WB Games Montréal"],
+        "genre": ["Fight", "Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/vi2o43IfXio",
+    },
+    {
+        "image": "../assets/NBA2K23.jpg",
+        "name": "NBA 2K23",
+        "opinion": 4.8,
+        "price": 27.25,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Nintendo Switch"],
+        "developers": ["Visual Concepts"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embed/MCU18y25WeQ",
+    },
+    {
+        "image": "../assets/mortalKombatUltimate.jpg",
+        "name": "MORTAL KOMBAT ULTIMATE",
+        "opinion": 4.5,
+        "price": 30.74,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Xbox Series"],
+        "developers": ["NetherRealm Studios", "Shiver Entertainment", "QLOC"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/OEQsQqkEKMA"
+    },
+    {
+        "image": "../assets/overCookedAllYouCanEat.jpg",
+        "name": "OVERCOOKED! All you can eat",
+        "opinion": 4.4,
+        "price": 26.39,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Ghost Town Games", "Team17"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embed/H-FR-apUaX0",
+    },
+    {
+        "image": "../assets/callOfDuttyVanguard.jpg",
+        "name": "CALL OF DUTTY: Vanguard",
+        "opinion": 4.3,
+        "price": 80.92,
+        "age": 18,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Infinity Ward"],
+        "genre": ["Shoter", "Open World"],
+        "trailer": "https://www.youtube.com/embed/zdVk4hHQTn0",
+    },
+    {
+        "image": "../assets/NBA2K21.jpg",
+        "name": "NBA 2K21",
+        "opinion": 4.4,
+        "price": 40.36,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Visual Concepts"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embed/RJ_bxl5DwkM",
+    },
+    {
+        "image": "../assets/atari50.jpg",
+        "name": "ATARI 50",
+        "opinion": 4.8,
+        "price": 40.11,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Xbox Series"],
+        "developers": ["Digital Eclipse"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embed/CQKwj_hi-Gc",
+    },
+    {
+        "image": "../assets/ninjaTurtles.jpg",
+        "name": "Teenage Mutant Ninja Turtles: The Cowabunga",
+        "opinion": 4.7,
+        "price": 40.33,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Xbox Series"],
+        "developers": ["Digital Eclipse"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embed/nlaZF5bxc50",
+    },
+    {
+        "image": "../assets/asterixAndObelix.jpg",
+        "name": "Asterix & Obelix Slap Them All !",
+        "opinion": 4.6,
+        "price": 29.84,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Mr. Nutz Studio"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embed/b-xZcfZMryQ",
+    },
+    {
+        "image": "../assets/cavemanNinja.jpg",
+        "name": "JOE and MAC Caveman Ninja",
+        "opinion": 4.1,
+        "price": 40.11,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Meridiem Games"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embed/Lkg1IBf__kk",
+    },
+    {
+        "image": "../assets/kaoTheKangaroo.jpg",
+        "name": "Kao The Kangaroo",
+        "opinion": 4.3,
+        "price": 40.11,
+        "age": 7,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Xbox Series"],
+        "developers": ["Tate Multimedia", "Titus Interactive"],
+        "genre": ["Platforms", "Open World", "Fight"],
+        "trailer": "https://www.youtube.com/embed/Ld9FVoL_5XI",
+    },
+    {
+        "image": "../assets/puyoPuyoTetris.jpg",
+        "name": "Puyo Puyo Tetris 2",
+        "opinion": 4.6,
+        "price": 39.83,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Xbox Series"],
+        "developers": ["Sonic Team", "Sega"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embed/Xbw6kgS_zlU",
+    },
+    {
+        "image": "../assets/destructionAllStars.jpg",
+        "name": "Destruction Allstars",
+        "opinion": 3.7,
+        "price": 19.84,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Lucid Games", "Xdev Software GmbH", "Wushu Studios Ltd"],
+        "genre": ["Racing", "Fight"],
+        "trailer": "https://www.youtube.com/embed/nNO1k5NCDOk",
+    },
+    {
+        "image": "../assets/monsterTruck.jpg",
+        "name": "Monster Truck Championship",
+        "opinion": 4.0,
+        "price": 19.99,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Nintendo Switch"],
+        "developers": ["Teyon"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embed/ChTSTqf4sL0",
+    },
+    {
+        "image": "../assets/edgeOfEternity.jpg",
+        "name": "Edge of Eternity",
+        "opinion": 4.5,
+        "price": 38.52,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Midgar Studio", "Ubitus K.K."],
+        "genre": ["Open World", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/5_k2DRoZ8fA",
+    },
+    {
+        "image": "../assets/w2c10.jpg",
+        "name": "World Rally Championship 10",
+        "opinion": 4.4,
+        "price": 29.35,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Kylotonn"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embed/RWzzZwNSLzM",
+    },
+    {
+        "image": "../assets/theDioFieldChronicle.jpg",
+        "name": "The DioField Chronicle",
+        "opinion": 4.4,
+        "price": 35.03,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Square Enix", "Lancarse", "Squaresoft"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/a_Jnap7qQK8",
+    },
+    {
+        "image": "../assets/inertialDrift.jpg",
+        "name": "Inertial Drift Twilight Rivals Edition",
+        "opinion": 4.7,
+        "price": 29.85,
+        "age": 3,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Level 91 Entertainment"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embed/PWwlfZsovuU",
+    },
+    {
+        "image": "../assets/tacticOgreReborn.jpg",
+        "name": "Tactics Ogre Reborn",
+        "opinion": 4.7,
+        "price": 53.26,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Square Enix", "Quest Corporation", "Riverhillsoft"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed/fFD3mWpFbqM",
+    },
+    {
+        "image": "../assets/gothamKnightsX.jpg",
+        "name": "GOTHAM KNIGTHS",
+        "opinion": 4.5,
+        "price": 29.72,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["QLOC", "WB Games Montréal"],
+        "genre": ["Fight", "Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedMawcZFHaa54",
+    },
+    {
+        "image": "../assets/NBA2K23X.jpg",
+        "name": "NBA 2K23",
+        "opinion": 4.7,
+        "price": 21.26,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Visual Concepts"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedxlVS-EvChlM",
+    },
+    {
+        "image": "../assets/callOfdutyX.jpg",
+        "name": "Call of Duty: Modern Warfare II",
+        "opinion": 4.6,
+        "price": 79.93,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Infinity Ward"],
+        "genre": ["Shoter"],
+        "trailer": "https://www.youtube.com/embed3wzfnmXK1tg",
+    },
+    {
+        "image": "../assets/theCalistoProtocolX.jpg",
+        "name": "The Calisto Protocols",
+        "opinion": 4.6,
+        "price": 69.95,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["PUBG Studios", "Striking distance studios"],
+        "genre": ["Horror", "Shoter", "Fight"],
+        "trailer": "https://www.youtube.com/embedEZZhzSgLfCM",
+    },
+    {
+        "image": "../assets/aPlaguetaleRequiemX.jpg",
+        "name": "A Plague Tale: Requiem",
+        "opinion": 4.8,
+        "price": 52.11,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Asobo Studio"],
+        "genre": ["Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embedtE0v23BL8dA",
+    },
+    {
+        "image": "../assets/personasRoyalX.jpg",
+        "name": "Persona 5 Royal",
+        "opinion": 4.6,
+        "price": 60.11,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Atlus, P Studio"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embed2smqR3An4s4",
+    },
+    {
+        "image": "../assets/personasRoyal.jpg",
+        "name": "Persona 5 Royal",
+        "opinion": 4.9,
+        "price": 52.52,
+        "age": 16,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Atlus, P Studio"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedzhIp0nVsNXc",
+    },
+    {
+        "image": "../assets/soulHackersX.jpg",
+        "name": "Soul Hackers 2",
+        "opinion": 4.5,
+        "price": 37.47,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Atlus"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedkrci4Cr-UcQ",
+    },
+    {
+        "image": "../assets/soulHackers.jpg",
+        "name": "Soul Hackers 2",
+        "opinion": 4.5,
+        "price": 35.49,
+        "age": 16,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Atlus"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedVNP9GFsz82k",
+    },
+    {
+        "image": "../assets/newTalesFromTheBorderlands.jpg",
+        "name": "New tales from the Borderlands",
+        "opinion": 4.0,
+        "price": 19.97,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Gearbox Software"],
+        "genre": ["Adventure"],
+        "trailer": "https://www.youtube.com/embedEg86enjOUDc",
+    },
+    {
+        "image": "../assets/theEternalCylinder.jpg",
+        "name": "The Eternal Cylinder",
+        "opinion": 4.7,
+        "price": 34.45,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["ACE Team"],
+        "genre": ["Open World"],
+        "trailer": "https://www.youtube.com/embedqpgOXYUNoNA",
+    },
+    {
+        "image": "../assets/ironHarvestX.jpg",
+        "name": "Iron Harvest",
+        "opinion": 4.3,
+        "price": 19.96,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["King Art Games"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embedTSdXBCAILew",
+    },
+    {
+        "image": "../assets/.jpg",
+        "name": "Forza Horizon 5",
+        "opinion": 4.7,
+        "price": 70.36,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Playground Games"],
+        "genre": ["Racing", "Adventure"],
+        "trailer": "https://www.youtube.com/embedRv7xLt5yNsM",
+    },
+    {
+        "image": "../assets/haloInfinity.jpg",
+        "name": "Halo Infinity",
+        "opinion": 4.6,
+        "price": 34.93,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["343 Industries"],
+        "genre": ["Shoter"],
+        "trailer": "https://www.youtube.com/embedsUs2Vb215-w",
+    },
+    {
+        "image": "../assets/deadIsland.jpg",
+        "name": "Dead Island 2",
+        "opinion": 4.1,
+        "price": 76.16,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Deep Silver Dambuster Studios"],
+        "genre": ["Horror", "Adventure"],
+        "trailer": "https://www.youtube.com/embedyqvnKCqgGH0",
+    },
+    {
+        "image": "../assets/atomicHeart.jpg",
+        "name": "Atomic Heart",
+        "opinion": 4.1,
+        "price": 72.06,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Mundfish"],
+        "genre": ["Shoter"],
+        "trailer": "https://www.youtube.com/embed5ggo94s1wx4",
+    },
+    {
+        "image": "../assets/redFall.jpg",
+        "name": "Red Fall",
+        "opinion": 4.2,
+        "price": 74.74,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Arkane Studios"],
+        "genre": ["Shoter", "Horror"],
+        "trailer": "https://www.youtube.com/embedtdrRQo-ZRUw",
+    },
+    {
+        "image": "../assets/howartsLegacyX.jpg",
+        "name": "Howarts Legacy",
+        "opinion": 4.8,
+        "price": 72.27,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Portkey Games"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedBtyBjOW8sGY",
+    },
+    {
+        "image": "../assets/howartsLegacyS.jpg",
+        "name": "Howarts Legacy",
+        "opinion": 4.9,
+        "price": 61.88,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Play Station 5"],
+        "developers": ["Portkey Games"],
+        "genre": ["Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedNthGfn_ddRQ",
+    },
+    {
+        "image": "../assets/noMoreHeroesIII.jpg",
+        "name": "No More Heroes III",
+        "opinion": 3.8,
+        "price": 60.33,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Grasshopper Manufacture"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedWdG8VmAJZ98",
+    },
+    {
+        "image": "../assets/saintsRow.jpg",
+        "name": "Saints Row",
+        "opinion": 3.9,
+        "price": 31.46,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Volition Inc.", "Sperasoft"],
+        "genre": ["Racing", "Shoter", "Open World"],
+        "trailer": "https://www.youtube.com/embedLjtwFdBM9Zo",
+    },
+    {
+        "image": "../assets/noMoreHeroesIIIS.jpg",
+        "name": "No More Heroes III",
+        "opinion": 4.6,
+        "price": 48.36,
+        "age": 18,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Grasshopper Manufacture"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embed7TIogRoa3A8",
+    },
+    {
+        "image": "../assets/scarsAboveX.jpg",
+        "name": "Scars Above",
+        "opinion": 4.1,
+        "price": 38.19,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Koch Media GmbH"],
+        "genre": ["Shoter", "Adventure"],
+        "trailer": "https://www.youtube.com/embed88VqoDdiB98",
+    },
+    {
+        "image": "../assets/bloodBowl3.jpg",
+        "name": "Blood Bowl 3",
+        "opinion": 3.8,
+        "price": 41.32,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Cyanide Studio"],
+        "genre": ["Sports", "Strategy"],
+        "trailer": "https://www.youtube.com/embedxSvH3Y4hjys",
+    },
+    {
+        "image": "../assets/brambleTheMountainKing.jpg",
+        "name": "Bramble: The Mountain King",
+        "opinion": 4.6,
+        "price": 38.44,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Dimfrost Studio"],
+        "genre": ["Horror", "Platforms", "Adventure"],
+        "trailer": "https://www.youtube.com/embedN4rvPNYB4z0",
+    },
+    {
+        "image": "../assets/afterimage.jpg",
+        "name": "Afterimage",
+        "opinion": 4.4,
+        "price": 34.26,
+        "age": 7,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Aurogon Shanghai"],
+        "genre": ["Platforms", "Adventure"],
+        "trailer": "https://www.youtube.com/embedVtPE13aenfU",
+    },
+    {
+        "image": "../assets/afterimageS.jpg",
+        "name": "Afterimage",
+        "opinion": 4.2,
+        "price": 32.52,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series"],
+        "developers": ["Aurogon Shanghai"],
+        "genre": ["Platforms", "Adventure"],
+        "trailer": "https://www.youtube.com/embed_msrs53rNJk",
+    },
+    {
+        "image": "../assets/clashArtifactsOfChaos.jpg",
+        "name": "Clash Artifacts of Chaos",
+        "opinion": 4.1,
+        "price": 40.22,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["ACE Team"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedB0VSx4OPO_0",
+    },
+    {
+        "image": "../assets/wantedDead.jpg",
+        "name": "Wanted: Dead",
+        "opinion": 3.5,
+        "price": 51.23,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Soleil"],
+        "genre": ["Rol", "Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedJFUb6w9ZrKw",
+    },
+    {
+        "image": "../assets/sonicFrontiersX.jpg",
+        "name": "Sonic Frontiers",
+        "opinion": 4.6,
+        "price": 62.34,
+        "age": 7,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Sonic Team"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embedlpfB04QtG9w",
+    },
+    {
+        "image": "../assets/minecraftLegendsX.jpg",
+        "name": "Minecraft Leyends",
+        "opinion": 4.4,
+        "price": 39.15,
+        "age": 7,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Mojang Studios"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embedrsrtI0fpkV4",
+    },
+    {
+        "image": "../assets/minecraftLegendsS.jpg",
+        "name": "Minecraft Leyends",
+        "opinion": 4.0,
+        "price": 50.23,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["Mojang Studios"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embedOBlv7Cdydeg",
+    },
+    {
+        "image": "../assets/nickelodeonAllStarBrawlX.jpg",
+        "name": "Nickelodeon All Star Brawl",
+        "opinion": 4.5,
+        "price": 22.67,
+        "age": 7,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Ludosity"],
+        "genre": ["Fight"],
+        "trailer": "https://www.youtube.com/embedSfJy5kwKJvg",
+    },
+    {
+        "image": "../assets/nickelodeonAllStarBrawlS.jpg",
+        "name": "Nickelodeon All Star Brawl",
+        "opinion": 4.6,
+        "price": 33.28,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["Ludosity"],
+        "genre": ["Fight"],
+        "trailer": "https://www.youtube.com/embedBIuGYhVDAis",
+    },
+    {
+        "image": "../assets/metroExodusX.jpg",
+        "name": "Metro Exodus",
+        "opinion": 4.4,
+        "price": 44.97,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["4A Games"],
+        "genre": ["Shoter", "Horror"],
+        "trailer": "https://www.youtube.com/embeduaBfkITlAQ0",
+    },
+    {
+        "image": "../assets/likeADragonIshinX.jpg",
+        "name": "Like A Dragon: ISHIN",
+        "opinion": 4.7,
+        "price": 62.10,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Ryu ga Gotoku Studio"],
+        "genre": ["Fight", "Adventure", "Open World"],
+        "trailer": "https://www.youtube.com/embedbTpMuSLx6Bs",
+    },
+    {
+        "image": "../assets/wolongFallenDinastyX.jpg",
+        "name": "WO LONG Fallen Dinasty",
+        "opinion": 3.9,
+        "price": 75.29,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Team Ninja"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedPrCkv47t9c4",
+    },
+    {
+        "image": "../assets/aeternaNoctisS.jpg",
+        "name": "Aeterna Noctis",
+        "opinion": 4.8,
+        "price": 27.62,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Aeternum Game Studio"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embedzR0n4a75o2o",
+    },
+    {
+        "image": "../assets/astroniteS.jpg",
+        "name": "Astronite",
+        "opinion": 4.3,
+        "price": 36.08,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Dume Games Studio"],
+        "genre": ["Platforms", "Adventure", "Shoter"],
+        "trailer": "https://www.youtube.com/embedmD__P1Vy_7E",
+    },
+    {
+        "image": "../assets/marioKart.jpg",
+        "name": "Mario Kart Deluxe",
+        "opinion": 4.8,
+        "price": 46.82,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Nintendo Entertainment Planning & Development"],
+        "genre": ["Racing", "Classic"],
+        "trailer": "https://www.youtube.com/embedN_Hw_M5ENCQ",
+    },
+    {
+        "image": "../assets/dnfDuel.jpg",
+        "name": "DNF Duel",
+        "opinion": 4.6,
+        "price": 53.14,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Arc System Works", "Neople", "8ing/Raizing"],
+        "genre": ["Fight"],
+        "trailer": "https://www.youtube.com/embed5Puyrh7xoAU",
+    },
+    {
+        "image": "../assets/dnfDuelP.jpg",
+        "name": "DNF Duel",
+        "opinion": 4.7,
+        "price": 48.89,
+        "age": 12,
+        "console": "Play Station 5",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch"],
+        "developers": ["Arc System Works", "Neople", "8ing/Raizing"],
+        "genre": ["Fight"],
+        "trailer": "https://www.youtube.com/embedNDiDr-h6SqQ",
+    },
+    {
+        "image": "../assets/judgment.jpg",
+        "name": "Judgment",
+        "opinion": 4.8,
+        "price": 52.69,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Ryu ga Gotoku Studio", "Sega"],
+        "genre": ["Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embedlepClxz8kEM",
+    },
+    {
+        "image": "../assets/octopathTraveler.jpg",
+        "name": "Octopath Traveler",
+        "opinion": 4.7,
+        "price": 53.17,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Square Enix", "Acquire"],
+        "genre": ["Rol"],
+        "trailer": "https://www.youtube.com/embedFmi8KrntszI",
+    },
+    {
+        "image": "../assets/zeldaTearsKingdom.jpg",
+        "name": "Zelda Tears of the Kingdom",
+        "opinion": 5.0,
+        "price": 68.34,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Nintendo Entertainment Planning & Development"],
+        "genre": ["Adventure", "Open World", "Fight", "Strategy"],
+        "trailer": "https://www.youtube.com/embedsjxLF4IYnJc",
+    },
+    {
+        "image": "../assets/inked.jpg",
+        "name": "Inked: A Tale of Love",
+        "opinion": 4.4,
+        "price": 34.61,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Nuverse", "Starbreeze Studios", "Pixmain Pte. Ltd."],
+        "genre": ["Adventure", "Platforms", "Strategy"],
+        "trailer": "https://www.youtube.com/embed6xPw6elq4dA",
+    },
+    {
+        "image": "../assets/fairyFencerFRefrainChordS.jpg",
+        "name": "Fairy Fencer F: Refrain Chord",
+        "opinion": 4.2,
+        "price": 53.19,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Sting Entertainment", "Compile Heart", "Idea Factory"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embedvCfZcp_QQHY",
+    },
+    {
+        "image": "../assets/bigBrain.jpg",
+        "name": "Big Brain Academy",
+        "opinion": 4.6,
+        "price": 32.59,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Nintendo Entertainment Planning & Development"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embediIKi_DUJwPY",
+    },
+    {
+        "image": "../assets/donkeyKongCountry.jpg",
+        "name": "Donkey Kong Country",
+        "opinion": 4.8,
+        "price": 90.15,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Retro Studios"],
+        "genre": ["Platforms", "Adventure", "Classic"],
+        "trailer": "https://www.youtube.com/embedOr5HB1Lo8b0",
+    },
+    {
+        "image": "../assets/yoshisWorld.jpg",
+        "name": "Yoshi`s World",
+        "opinion": 4.7,
+        "price": 56.43,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Good-feel"],
+        "genre": ["Platforms", "Adventure", "Classic"],
+        "trailer": "https://www.youtube.com/embedQsFJbJccEkI",
+    },
+    {
+        "image": "../assets/cryMachina.jpg",
+        "name": "Cry Machina",
+        "opinion": 4.8,
+        "price": 61.16,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["FuRyu"],
+        "genre": ["Rol", "Shoter", "Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedLC2qz3QKxjQ",
+    },
+    {
+        "image": "../assets/fiveNightsAtFreddysSecurityBreachS.jpg",
+        "name": "Five Nights at Freddy's: Security Breach",
+        "opinion": 4.5,
+        "price": 40.52,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Steel Wool Studios"],
+        "genre": ["Adventure", "Horror"],
+        "trailer": "https://www.youtube.com/embed4cKB0r-NDO8",
+    },
+    {
+        "image": "../assets/hoa.jpg",
+        "name": "Hoa",
+        "opinion": 4.7,
+        "price": 38.41,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Skrollcat Studio"],
+        "genre": ["Platforms", "Strategy"],
+        "trailer": "https://www.youtube.com/embedN_JvdzwO92k",
+    },
+    {
+        "image": "../assets/signalis.jpg",
+        "name": "Signalis",
+        "opinion": 4.4,
+        "price": 28.59,
+        "age": 16,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["rose-engine"],
+        "genre": ["Horror"],
+        "trailer": "https://www.youtube.com/embedsk6HJVC3bpk",
+    },
+    {
+        "image": "../assets/swordOfTheVagrant.jpg",
+        "name": "Sword of the Vagrant",
+        "opinion": 4.2,
+        "price": 41.23,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["O.T.K Games", "DICO Co. Ltd"],
+        "genre": ["Platforms", "Fight", "Horror", "Rol"],
+        "trailer": "https://www.youtube.com/embedGkXlsTNRJnQ",
+    },
+    {
+        "image": "../assets/grandTheftAuto.jpg",
+        "name": "Grand Theft Auto V",
+        "opinion": 4.6,
+        "price": 36.33,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Rock Start Games"],
+        "genre": ["Open World", "Shoter"],
+        "trailer": "https://www.youtube.com/embed0wIzQC54ySk",
+    },
+    {
+        "image": "../assets/deliverUsTheMoonS.jpg",
+        "name": "DELIVER US THE MOON",
+        "opinion": 4.3,
+        "price": 27.47,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["KeokeN Interactive"],
+        "genre": ["Strategy", "Adventure", "Platforms"],
+        "trailer": "https://www.youtube.com/embedv_RPWDtnRe8",
+    },
+    {
+        "image": "../assets/deathsDoorS.jpg",
+        "name": "DEATH`S DOOR",
+        "opinion": 4.7,
+        "price": 28.49,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Acid Nerve"],
+        "genre": ["Rol", "Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embedIbPg9ErW0Q8",
+    },
+    {
+        "image": "../assets/catMaze.jpg",
+        "name": "Cat Maze",
+        "opinion": 4.6,
+        "price": 42.71,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["RedBlack Spade", "Ratalaika Games"],
+        "genre": ["Rol", "Platforms"],
+        "trailer": "https://www.youtube.com/embedEQsOTPPhFPM",
+    },
+    {
+        "image": "../assets/vampireTheMasqueradeS.jpg",
+        "name": "VAMPIRE: The masquerade",
+        "opinion": 4.1,
+        "price": 38.63,
+        "age": 16,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["HardSuit Labs", "Paradox Interactive"],
+        "genre": ["Rol", "Horror", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embediOrTyqrSHMo",
+    },
+    {
+        "image": "../assets/steelRisingX.jpg",
+        "name": "STEEL RISING",
+        "opinion": 4.3,
+        "price": 57.77,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Spiders"],
+        "genre": ["Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embed7WmeMQRF00s",
+    },
+    {
+        "image": "../assets/mortalKombatUltimateS.jpg",
+        "name": "MORTAL KOMBAT ULTIMATE",
+        "opinion": 4.1,
+        "price": 34.19,
+        "age": 18,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["NetherRealm Studios", "Shiver Entertainment", "QLOC"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedV7A4sF5qLsU"
+    },
+    {
+        "image": "../assets/overCookedAllYouCanEatS.jpg",
+        "name": "OVERCOOKED! All you can eat",
+        "opinion": 4.6,
+        "price": 40.32,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Ghost Town Games", "Team17"],
+        "genre": ["Platforms"],
+        "trailer": "https://www.youtube.com/embedawQkrzqfjUg",
+    },
+    {
+        "image": "../assets/callOfDuttyVanguardX.jpg",
+        "name": "CALL OF DUTTY: Vanguard",
+        "opinion": 4.5,
+        "price": 75.57,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Infinity Ward"],
+        "genre": ["Shoter", "Open World"],
+        "trailer": "https://www.youtube.com/embedEnoX0eJUZ04",
+    },
+    {
+        "image": "../assets/NBA2K21X.jpg",
+        "name": "NBA 2K21",
+        "opinion": 4.3,
+        "price": 48.46,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Visual Concepts"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedh9JvKAaJCvI",
+    },
+    {
+        "image": "../assets/tokyo2020.jpg",
+        "name": "Tokyo 2020",
+        "opinion": 4.5,
+        "price": 43.32,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Visual Concepts"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedb74Z-bO2H0o",
+    },
+    {
+        "image": "../assets/NBA2K23S.jpg",
+        "name": "NBA 2K23",
+        "opinion": 4.7,
+        "price": 21.26,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["Visual Concepts"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedZcGDa2XDIWI",
+    },
+    {
+        "image": "../assets/atari50S.jpg",
+        "name": "ATARI 50",
+        "opinion": 4.8,
+        "price": 38.98,
+        "age": 16,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["Digital Eclipse"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embedJxbtZkIx4Tg",
+    },
+    {
+        "image": "../assets/ninjaTurtlesS.jpg",
+        "name": "Teenage Mutant Ninja Turtles: The Cowabunga",
+        "opinion": 4.8,
+        "price": 38.79,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["Digital Eclipse"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embeduJVkKlnKXME",
+    },
+    {
+        "image": "../assets/atari50X.jpg",
+        "name": "ATARI 50",
+        "opinion": 4.8,
+        "price": 41.50,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Play Station 5"],
+        "developers": ["Digital Eclipse"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embedk0AXmxaHwbU",
+    },
+    {
+        "image": "../assets/ninjaTurtlesX.jpg",
+        "name": "Teenage Mutant Ninja Turtles: The Cowabunga",
+        "opinion": 4.8,
+        "price": 39.97,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Digital Eclipse"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embedtfLRsejZA64",
+    },
+    {
+        "image": "../assets/tacticOgreRebornS.jpg",
+        "name": "Tactics Ogre Reborn",
+        "opinion": 4.7,
+        "price": 51.50,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Square Enix", "Quest Corporation", "Riverhillsoft"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embedqA9z8-AOmi0",
+    },
+    {
+        "image": "../assets/reverieKnights.jpg",
+        "name": "Reverie Knights Tactics",
+        "opinion": 4.5,
+        "price": 27.99,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["40 Giants Entertainment"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed8L9Kg1WTzb4",
+    },
+    {
+        "image": "../assets/kaoTheKangarooX.jpg",
+        "name": "Kao The Kangaroo",
+        "opinion": 4.5,
+        "price": 37.12,
+        "age": 7,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Nintendo Switch", "Play Station 5"],
+        "developers": ["Tate Multimedia", "Titus Interactive"],
+        "genre": ["Platforms", "Open World", "Fight"],
+        "trailer": "https://www.youtube.com/embeduQnbqibVQPo",
+    },
+    {
+        "image": "../assets/kaoTheKangarooS.jpg",
+        "name": "Kao The Kangaroo",
+        "opinion": 4.3,
+        "price": 35.13,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Xbox Series"],
+        "developers": ["Tate Multimedia", "Titus Interactive"],
+        "genre": ["Platforms", "Open World", "Fight"],
+        "trailer": "https://www.youtube.com/embedOSf6n7lDlck",
+    },
+    {
+        "image": "../assets/puyoPuyoTetrisS.jpg",
+        "name": "Puyo Puyo Tetris 2",
+        "opinion": 4.4,
+        "price": 37.83,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Play Station 5"],
+        "developers": ["Sonic Team", "Sega"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embedqblxiyxgKX0",
+    },
+    {
+        "image": "../assets/puyoPuyoTetrisX.jpg",
+        "name": "Puyo Puyo Tetris 2",
+        "opinion": 4.5,
+        "price": 35.47,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Sonic Team", "Sega"],
+        "genre": ["Strategy"],
+        "trailer": "https://www.youtube.com/embed_-sPtVEw6xQ",
+    },
+    {
+        "image": "../assets/blasphemous.jpg",
+        "name": "Blasphemous",
+        "opinion": 4.7,
+        "price": 52.53,
+        "age": 16,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["The Game Kitchen,"],
+        "genre": ["Platforms", "Fight", "Horror", "Strategy"],
+        "trailer": "https://www.youtube.com/embed9DePPiw44is",
+    },
+    {
+        "image": "../assets/taleOfShadow.jpg",
+        "name": "Tandem : A Tale of Shadow",
+        "opinion": 4.8,
+        "price": 40.52,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Monochrome"],
+        "genre": ["Platforms", "Strategy", "Horror"],
+        "trailer": "https://www.youtube.com/embedC3GWHwVNaiU",
+    },
+    {
+        "image": "../assets/littleNightmares.jpg",
+        "name": "Little Nightmares",
+        "opinion": 4.8,
+        "price": 37.49,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Supermassive Games", "Tarsier Studios"],
+        "genre": ["Strategy", "Horror", "Platforms"],
+        "trailer": "https://www.youtube.com/embedDD_t503qMXc",
+    },
+    {
+        "image": "../assets/trekToYomi.jpg",
+        "name": "Trek to Yomi",
+        "opinion": 4.3,
+        "price": 30.24,
+        "age": 18,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Flying Wild Hog", "Leonard Menchiari", "TFL Studios LLC"],
+        "genre": ["Open World", "Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embeddzHg870vglo",
+    },
+    {
+        "image": "../assets/pokemonLegendsArceus.jpg",
+        "name": "pokemon Legends Arceus",
+        "opinion": 4.8,
+        "price": 51.84,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Game Freak"],
+        "genre": ["Open World", "Adventure", "Strategy", "Rol"],
+        "trailer": "https://www.youtube.com/embed9RbjFjYgNNg&t=254s",
+    },
+    {
+        "image": "../assets/zeldaBreathOfTheWild.jpg",
+        "name": "Zelda Breath of the Wild",
+        "opinion": 4.8,
+        "price": 65.36,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Nintendo Entertainment Planning & Development"],
+        "genre": ["Open World", "Adventure", "Fight", "Strategy"],
+        "trailer": "https://www.youtube.com/embedofH5ptn5w-A",
+    },
+    {
+        "image": "../assets/cavemanNinjaS.jpg",
+        "name": "JOE and MAC Caveman Ninja",
+        "opinion": 3.8,
+        "price": 38.64,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Meridiem Games"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embedVEtrfSIQCPQ",
+    },
+    {
+        "image": "../assets/asterixAndObelixS.jpg",
+        "name": "Asterix & Obelix Slap Them All !",
+        "opinion": 4.5,
+        "price": 27.80,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Mr. Nutz Studio"],
+        "genre": ["Classic", "Platforms", "Fight"],
+        "trailer": "https://www.youtube.com/embedW0tlF-RP9Jo",
+    },
+    {
+        "image": "../assets/theDioFieldChronicleX.jpg",
+        "name": "The DioField Chronicle",
+        "opinion": 4.7,
+        "price": 37.39,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Square Enix", "Lancarse", "Squaresoft"],
+        "genre": ["Rol", "Strategy", "Adventure"],
+        "trailer": "https://www.youtube.com/embed0jeMhdC9ajk",
+    },
+    {
+        "image": "../assets/inertialDriftS.jpg",
+        "name": "Inertial Drift Twilight Rivals Edition",
+        "opinion": 4.0,
+        "price": 27.10,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Level 91 Entertainment"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embedQCsixzJ_1yA",
+    },
+    {
+        "image": "../assets/w2c10X.jpg",
+        "name": "World Rally Championship 10",
+        "opinion": 4.4,
+        "price": 35.22,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5"],
+        "developers": ["Kylotonn"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embed_zlqLYS2f70",
+    },
+    {
+        "image": "../assets/monsterTruckS.jpg",
+        "name": "Monster Truck Championship",
+        "opinion": 3.9,
+        "price": 18.75,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": false,
+        "sharedWith": ["Xbox Series", "Play Station 5"],
+        "developers": ["Teyon"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embedpqpZh9E9B_Y",
+    },
+    {
+        "image": "../assets/monsterTruckX.jpg",
+        "name": "Monster Truck Championship",
+        "opinion": 4.1,
+        "price": 20.35,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["Teyon"],
+        "genre": ["Racing"],
+        "trailer": "https://www.youtube.com/embedohDkkLotw9E",
+    },
+    {
+        "image": "../assets/marioStrikers.jpg",
+        "name": "Mario Strikers Football League",
+        "opinion": 4.8,
+        "price": 54.77,
+        "age": 7,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Next Level Games"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedfkcbEe_Hox4",
+    },
+    {
+        "image": "../assets/PGA2K23.jpg",
+        "name": "PGA 2K23",
+        "opinion": 4.4,
+        "price": 41.78,
+        "age": 3,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["HB Studios"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedtgTgzWFVtS0",
+    },
+    {
+        "image": "../assets/rainbowXS.jpg",
+        "name": "Rainbow SIX",
+        "opinion": 4.2,
+        "price": 15.38,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Ubisoft"],
+        "genre": ["Shoter", "Strategy"],
+        "trailer": "https://www.youtube.com/embedZH0VC-MTAzE",
+    },
+    {
+        "image": "../assets/theKingOfFighters.jpg",
+        "name": "The King Of Fighters",
+        "opinion": 4.8,
+        "price": 35.54,
+        "age": 12,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["SNK Eolith"],
+        "genre": ["Fight"],
+        "trailer": "https://www.youtube.com/embedEBj93vo8fOI",
+    },
+    {
+        "image": "../assets/mortalKombatUltimateX.jpg",
+        "name": "MORTAL KOMBAT ULTIMATE",
+        "opinion": 4.7,
+        "price": 29.09,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": false,
+        "sharedWith": ["Play Station 5", "Nintendo Switch"],
+        "developers": ["NetherRealm Studios", "Shiver Entertainment", "QLOC"],
+        "genre": ["Fight", "Adventure"],
+        "trailer": "https://www.youtube.com/embedvhjoNZIogRA"
+    },
+    {
+        "image": "../assets/thymesia.jpg",
+        "name": "Thymesia",
+        "opinion": 4.8,
+        "price": 32.37,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["OverBorder Studio"],
+        "genre": ["Open World", "Fight", "Adventure", "Horror"],
+        "trailer": "https://www.youtube.com/embedqJdw31GrUO4",
+    },
+    {
+        "image": "../assets/discoElysium.jpg",
+        "name": "Disco Elysium the Final Cut",
+        "opinion": 4.6,
+        "price": 42.32,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["ZA/UM"],
+        "genre": ["Open World", "Rol", "Adventure"],
+        "trailer": "https://www.youtube.com/embedYODkDdDkdps",
+    },
+    {
+        "image": "../assets/gunGrave.jpg",
+        "name": "Gungrave G.O.R.E.",
+        "opinion": 4.7,
+        "price": 60.34,
+        "age": 16,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Iggymob"],
+        "genre": ["Shoter", "Adventure"],
+        "trailer": "https://www.youtube.com/embedj_uP2uvBvko",
+    },
+    {
+        "image": "../assets/moonscars.jpg",
+        "name": "Moonscars",
+        "opinion": 4.7,
+        "price": 35.78,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Black Mermaid"],
+        "genre": ["Platforms", "Horror", "Fight"],
+        "trailer": "https://www.youtube.com/embedNXR4xhukd7A",
+    },
+    {
+        "image": "../assets/outRaidersWorldSlayer.jpg",
+        "name": "Outriders Worldslayer",
+        "opinion": 4.7,
+        "price": 20.00,
+        "age": 18,
+        "console": "Xbox Series",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Square Enix"],
+        "genre": ["Rol", "Shoter", "Adventure"],
+        "trailer": "https://www.youtube.com/embedBSH1c9Pugq8",
+    },
+    {
+        "image": "../assets/jitsuSquad.jpg",
+        "name": "Jitsu Squad",
+        "opinion": 4.7,
+        "price": 28.89,
+        "age": 16,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Tanuki Creative Studio"],
+        "genre": ["Fight", "Platforms", "Classic"],
+        "trailer": "https://www.youtube.com/embedosve8RwMyog",
+    },
+    {
+        "image": "../assets/legendaryEleven.jpg",
+        "name": "Legendary Eleven",
+        "opinion": 3.7,
+        "price": 24.45,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Eclipse Games"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embedBoTMNbh5M3A",
+    },
+    {
+        "image": "../assets/snookers.jpg",
+        "name": "Snookers",
+        "opinion": 4.3,
+        "price": 34.65,
+        "age": 3,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Lab42"],
+        "genre": ["Sports"],
+        "trailer": "https://www.youtube.com/embed2VtPyBMmeI8",
+    },
+    {
+        "image": "../assets/segaMegaDriveClassics.jpg",
+        "name": "SEGA Mega Drive Classics",
+        "opinion": 4.5,
+        "price": 28.71,
+        "age": 12,
+        "console": "Nintendo Switch",
+        "exclusiveness": true,
+        "sharedWith": [],
+        "developers": ["Sega"],
+        "genre": ["Classic", "Platforms", "Rol", "Adventure", "Fight"],
+        "trailer": "https://www.youtube.com/embedOKs1sYh_nMQ",
+    },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+    // {
+    //     "image": "../assets/.jpg",
+    //     "name": "",
+    //     "opinion": ,
+    //     "price": ,
+    //     "age": ,
+    //     "console": "Xbox Series",
+    //     "exclusiveness": false,
+    //     "sharedWith": [],
+    //     "developers": [""],
+    //     "genre": ["", ""],
+    //     "trailer": "",
+    // },
+
+]
+
+module.exports = {
+    products,
+    providers
+}
